End of Chapter 6.0

+++ Math.ExpressionEvaluator/AddOperator.cs
@@ -7,7 +7,7 @@
             Precedence = 1 + precedenceBoost;
         }
         
-        public override int Compute(Operand left, Operand right)
+        public override double Compute(Operand left, Operand right)
         {
             return left.Value + right.Value;
         }
