+++ Math.ExpressionEvaluator/OperandFactory.cs
@@ -2,7 +2,7 @@
 {
     public class OperandFactory : IOperandFactory
     {
-        public Operand Create(int value)
+        public Operand Create(double value)
         {
             return new Operand(value);
         }
