+++ Math.ExpressionEvaluator/Operator.cs
@@ -3,6 +3,6 @@
     public abstract class Operator : Element
     {
         public int Precedence { get; protected set; }
-        public abstract int Compute(Operand left, Operand right);
+        public abstract double Compute(Operand left, Operand right);
     }
 }
