+++ Math.ExpressionEvaluator/SubOperator.cs
@@ -8,7 +8,7 @@
             Precedence = 1 + precedenceBoost;
         }
 
-        public override int Compute(Operand left, Operand right)
+        public override double Compute(Operand left, Operand right)
         {
             return left.Value - right.Value;
         }
